Add MusicXML parser
