+++ src/main/java/com/durangoretro/rescomp/ImageGenerator.java
@@ -78,7 +78,7 @@ public class ImageGenerator {
 				byte pixels[] = generateSpriteData(image, col, row, spriteWidth, spriteHeight);
 				resourceNameSB.setLength(0);
 				resourceNameSB.append(resourceName).append('_').append(row).append('_').append(col);
-				sb.append(getHexString(resourceNameSB.toString(), pixels)).append("\n");
+				sb.append(Util.getHexString(resourceNameSB.toString(), pixels)).append("\n");
 			}
 		}
 		return sb.toString();
@@ -103,34 +103,7 @@ public class ImageGenerator {
 		for(Byte b : bytes) {
 			pixels[i++]=b;
 		}
-		sb.append(getHexString(resourceNameSB.toString(), pixels)).append("\n");
+		sb.append(Util.getHexString(resourceNameSB.toString(), pixels)).append("\n");
 		return sb.toString();
 	}
-			
-	public static String getHexString(String resurceName, byte[] pixels) {
-		StringBuilder sb = new StringBuilder(5000);
-		sb.append("const unsigned char ").append(resurceName).append('[').append(pixels.length).append("] = {").append("\n");
-		for(int i=0; i<pixels.length; i++) {
-			sb.append("0x").append(String.format("%02X",pixels[i])).append(',');
-			if(i%32==31) {
-				sb.append("\n");
-			}
-		}
-		sb.append("\n").append("};");
-		return sb.toString();
-	}
-	
-	
-
-	/*
- const unsigned char tiles[8192] = {
-0x11,0x44,0x44,0x44,0x11,0x41,0x44,0x44,0x11,0x14,0x44,0x44,0x11,0x11,0x44,0x44,0x41,0x11,0x14,0x44,0x41,0x11,0x14,0x44,0x94,0x44,0x44,0x44,0x94,0x44,0x44,0x44,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFF,0xFF,0xFF,0xFF,0x00,0x00,0x00,0x00,0xFF,0xFF,0xFF,0xFF,
-};
-
-*/
-	
-		
-	
 }
