+++ src/main/java/com/durangoretro/rescomp/MusicGenerator.java
@@ -1,0 +1,15 @@
+package com.durangoretro.rescomp;
+
+import java.io.File;
+
+public class MusicGenerator {
+
+	public static byte[] convertToDurango(File file) {
+		
+		
+		
+		byte pixels[] = new byte[10];
+		return pixels;
+	}
+
+}
